// src/lib/notes/types/ErrorMessage.ts
export enum ErrorMessage {
  GRAPH_NOT_FOUND = "GRAPH_NOT_FOUND",
  NOTE_NOT_FOUND = "NOTE_NOT_FOUND",
  PINNED_NOTES_NOT_FOUND = "PINNED_NOTES_NOT_FOUND",
  PINNED_NOTE_NOT_FOUND = "PINNED_NOTE_NOT_FOUND",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
  INVALID_MIME_TYPE = "INVALID_MIME_TYPE",
  SLUG_EXISTS = "SLUG_EXISTS",
  INVALID_NOTE_STRUCTURE = "INVALID_NOTE_STRUCTURE",
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_FILENAME_EXTENSION = "INVALID_FILENAME_EXTENSION",
  NOT_SUPPORTED_BY_STORAGE_PROVIDER = "NOT_SUPPORTED_BY_STORAGE_PROVIDER",
  INVALID_SLUG = "INVALID_SLUG",
  INVALID_ALIAS = "INVALID_ALIAS",
  ALIAS_EXISTS = "ALIAS_EXISTS",
}